Move tests to blitz-tests
