+++ packages/blitz-dom/Cargo.toml
@@ -94,9 +94,6 @@ thread_local = { workspace = true }
 # WASM
 web-time = { workspace = true }
 
-[dev-dependencies]
-test-that = { workspace = true }
-
 # HACK: Blitz doesn't need to depend on objc2 directly. But this feature flag is necessary
 # to prevent debug builds from panicking.
 [target.'cfg(any(target_vendor = "apple"))'.dependencies]
