+++ tests/blitz-tests/Cargo.toml
@@ -31,6 +31,7 @@ accesskit = { workspace = true }
 markup5ever = { workspace = true }
 keyboard-types = { workspace = true }
 taffy = { workspace = true }
+test-that = { workspace = true }
 usvg = { workspace = true }
 
 [lib]
